requirements.txt
# фреймворк
Django>=4.0.7,<4.1.0
djangorestframework>=3.14.0,<3.15.0
# генерация Swagger
drf-yasg>=1.21.4,<1.22.0
# работа с переменными окружения
django-environ>=0.9.0,<0.10.0
# отладчик для Django
django-debug-toolbar>=3.6.0,<3.7.0

# адаптер для подключения к PostgreSQL
psycopg2-binary>=2.9.3,<2.10.0

# документация
Sphinx>=5.1.1,<5.2.0
sphinx-rtd-theme>=1.0.0,<2.0.0

# статический анализ кода (линтеры)
isort>=5.10.1,<5.11.0
flake8-django>=1.1.5,<1.2.0
pylint-django>=2.5.3,<2.6.0
django-stubs[compatible-mypy]>=1.12.0,<1.13.0

# автоматическое форматирование кода
black>=22.8.0,<22.9.0

redis>=4.3.4,<4.4.0
# распределенная очередь задач
celery>=5.2.7,<5.3.0
django-celery-beat>=2.3.0,<2.4.0
# работа с HTTP-запросами
httpx>=0.23.0,<0.24.0
# DTO и валидцаия данных
pydantic>=1.10.2,<1.11.0
